[MIN-354] moving liquidation scenarios in integration tests
